make-dist CLI

+++ quickevent/make-dist.sh
@@ -1,13 +1,68 @@
-#!/bin/sh
+#!/bin/bash
 
+APP_VER=0.0.1
+APP_NAME=quickevent
+SRC_DIR=/home/fanda/proj/$APP_NAME
+QT_DIR=/home/fanda/programs/qt5/5.13.2/gcc_64
+WORK_DIR=/tmp/_distro
 #DISTRO_VER=1.0.2
 APP_IMAGE_TOOL=/home/fanda/programs/appimagetool-x86_64.AppImage
 
-if [ $1 ]; then
-	DISTRO_VER=$1
-fi
+help() {
+	echo "make-dist.sh"
+	echo "\t--app-name application name, ie: myapp"
+	echo "\t--app-version application version, ie: 1.0.0"
+	echo "\t--src-dir application source dir, where *.pro file lies, ie: /home/me/myapp"
+	echo "\t--qt-dir QT dir, ie: /home/me/qt5/5.13.1/gcc_64"
+	echo "\t--work-dir directory where build will be made, ie: /tmp/myapp"
+	echo "\t--no-clean do not rebuild whole project when set to 1"
+	exit 0
+}
 
-SRC_DIR=/home/fanda/proj/quickbox
+while [[ $# -gt 0 ]]
+do
+key="$1"
+# echo key: $key
+case $key in
+	--app-name)
+	APP_NAME="$2"
+	shift # past argument
+	shift # past value
+	;;
+	--app-version)
+	APP_VER="$2"
+	shift # past argument
+	shift # past value
+	;;
+	--qt-dir)
+	QT_DIR="$2"
+	shift # past argument
+	shift # past value
+	;;
+	--src-dir)
+	SRC_DIR="$2"
+	shift # past argument
+	shift # past value
+	;;
+	--work-dir)
+	WORK_DIR="$2"
+	shift # past argument
+	shift # past value
+	;;
+	--no-clean)
+	NO_CLEAN=1
+	shift # past value
+	;;
+	-h|--help)
+	shift # past value
+	help
+	;;
+	*)    # unknown option
+	echo "ignoring argument $1"
+	shift # past argument
+	;;
+esac
+done
 
 if [ -z $DISTRO_VER ]; then
     DISTRO_VER=`grep APP_VERSION $SRC_DIR/quickevent/app/quickevent/src/appversion.h | cut -d\" -f2`
@@ -15,22 +70,23 @@ if [ -z $DISTRO_VER ]; then
 	#exit 1
 fi
 
-WORK_DIR=/home/fanda/t/_distro
-
-#USE_SYSTEM_QT=1
+echo APP_VER: $APP_VER
+echo APP_NAME: $APP_NAME
+echo SRC_DIR: $SRC_DIR
+echo QT_DIR: $QT_DIR
+echo WORK_DIR: $WORK_DIR
+echo NO_CLEAN: $NO_CLEAN
 
 if [ -z $USE_SYSTEM_QT ]; then
-    QT_DIR=/home/fanda/programs/qt5/5.13.0/gcc_64
-    echo using $QT_DIR
     QT_LIB_DIR=$QT_DIR/lib
     QMAKE=$QT_DIR/bin/qmake
-    DISTRO_NAME=quickevent-$DISTRO_VER-linux64
+    DISTRO_NAME=$APP_NAME-$DISTRO_VER-linux64
 else
     echo using system QT
     QT_DIR=/usr/lib/i386-linux-gnu/qt5
     QT_LIB_DIR=/usr/lib/i386-linux-gnu
     QMAKE=/usr/bin/qmake
-    DISTRO_NAME=quickevent-$DISTRO_VER-linux32
+    DISTRO_NAME=$APP_NAME-$DISTRO_VER-linux32
 fi
 
 $QMAKE -v
@@ -40,11 +96,14 @@ DIST_DIR=$WORK_DIR/$DISTRO_NAME
 DIST_LIB_DIR=$DIST_DIR/lib
 DIST_BIN_DIR=$DIST_DIR/bin
 
-rm -r $BUILD_DIR
-mkdir -p $BUILD_DIR
+if [ -z $NO_CLEAN ]; then
+	echo removing directory $WORK_DIR
+	rm -r $BUILD_DIR
+	mkdir -p $BUILD_DIR
+fi
 
 cd $BUILD_DIR
-$QMAKE $SRC_DIR/quickbox.pro CONFIG+=release CONFIG+=force_debug_info CONFIG+=separate_debug_info -r -spec linux-g++
+$QMAKE $SRC_DIR/$APP_NAME.pro CONFIG+=release CONFIG+=force_debug_info CONFIG+=separate_debug_info -r -spec linux-g++
 make -j2
 if [ $? -ne 0 ]; then
   echo "Make Error" >&2
@@ -99,5 +158,5 @@ rsync -av --exclude '*.debug' $QT_DIR/qml/QtQuick.2/ $DIST_BIN_DIR/QtQuick.2
 
 tar -cvzf $WORK_DIR/$DISTRO_NAME.tgz  -C $WORK_DIR ./$DISTRO_NAME
 
-rsync -av $SRC_DIR/quickevent/distro/QuickEvent.AppDir/* $DIST_DIR/
-ARCH=x86_64 $APP_IMAGE_TOOL $DIST_DIR $WORK_DIR/quickevent-${DISTRO_VER}-x86_64.AppImage
+rsync -av $SRC_DIR/$APP_NAME/distro/QuickEvent.AppDir/* $DIST_DIR/
+ARCH=x86_64 $APP_IMAGE_TOOL $DIST_DIR $WORK_DIR/$APP_NAME-${APP_VER}-x86_64.AppImage
